refactor: consolidate fragmented unit test files into canonical {type}_resource_test.go
Merges 12 scattered test files into their natural per-resource homes
so each resource has a single test file for all its unit tests.

Merges:
  kaas_create_test.go + kaas_rich_read_test.go + kaas_properties_test.go
      -> kaas_resource_test.go
  subnet_read_rich_test.go + subnet_configvalidator_test.go + subnet_validator_routes_test.go
      -> subnet_resource_test.go
  securityrule_any_protocol_test.go + securityrule_update_test.go
      -> securityrule_resource_test.go
  project_create_rich_test.go
      -> project_resource_test.go
  schedulejob_read_test.go
      -> schedulejob_resource_test.go

Distributes database_dbaasuser_create_test.go across 10 files:
  database, dbaasuser, project, elasticip, securitygroup, blockstorage,
  backup, keypair, snapshot, kms, databasegrant resource test files

Moves TestPtrToString from coverage_simple_unit_test.go
      -> kaas_data_source_test.go (beside its implementation in kaas_data_source.go)

Closes #299
