--- internal/provider/database_dbaasuser_create_test.go
@@ -1,502 +1,0 @@
-package provider
-
-import (
-	"context"
-	"net/http"
-	"testing"
-	"time"
-)
-
-// databaseCreateSuccessJSON is a response for the Database Create endpoint
-// that includes a top-level "name" field so that DatabaseResponse.Name is
-// non-empty and the provider can set data.Id = "testdb".
-// The WaitForResourceActive checker then calls Get("testdb") rather than
-// Get("") which the SDK would reject.
-const databaseCreateSuccessJSON = `{"name":"testdb","status":{"state":"Active"}}`
-
-// dbaasUserCreateSuccessJSON provides a top-level "username" field so that
-// UserResponse.Username is non-empty, letting the provider set data.Id and
-// call Get("test-user") rather than Get("").
-const dbaasUserCreateSuccessJSON = `{"username":"test-user","status":{"state":"Active"}}`
-
-func databaseCreateSuccessHandler(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	if r.Method == http.MethodPost {
-		w.WriteHeader(http.StatusCreated)
-	} else {
-		w.WriteHeader(http.StatusOK)
-	}
-	w.Write([]byte(databaseCreateSuccessJSON)) //nolint:errcheck
-}
-
-func dbaasUserCreateSuccessHandler(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	if r.Method == http.MethodPost {
-		w.WriteHeader(http.StatusCreated)
-	} else {
-		w.WriteHeader(http.StatusOK)
-	}
-	w.Write([]byte(dbaasUserCreateSuccessJSON)) //nolint:errcheck
-}
-
-// TestDatabaseCreate_Success verifies that database Create() succeeds when the
-// API response includes the database name (used as the resource ID) so that
-// the WaitForResourceActive poll can call Get with the right ID.
-func TestDatabaseCreate_Success(t *testing.T) {
-	oldActivePoll := waitForActivePollInterval
-	waitForActivePollInterval = 1 * time.Millisecond
-	t.Cleanup(func() { waitForActivePollInterval = oldActivePoll })
-
-	ctx := context.Background()
-
-	_, mockClient := newMockArubaClient(t, databaseCreateSuccessHandler)
-
-	res := NewDatabaseResource()
-	configureResource(ctx, t, res, mockClient)
-
-	req, resp := resourceCreateReq(ctx, t, res)
-	res.Create(ctx, req, resp)
-
-	if resp.Diagnostics.HasError() {
-		t.Errorf("Database Create() reported error: %v", resp.Diagnostics)
-	}
-}
-
-// TestDBaaSUserCreate_Success verifies that DBaaS user Create() succeeds when
-// the API response includes the username so the WaitForResourceActive checker
-// can call Get with the right username.
-func TestDBaaSUserCreate_Success(t *testing.T) {
-	oldActivePoll := waitForActivePollInterval
-	waitForActivePollInterval = 1 * time.Millisecond
-	t.Cleanup(func() { waitForActivePollInterval = oldActivePoll })
-
-	ctx := context.Background()
-
-	_, mockClient := newMockArubaClient(t, dbaasUserCreateSuccessHandler)
-
-	res := NewDBaaSUserResource()
-	configureResource(ctx, t, res, mockClient)
-
-	req, resp := resourceCreateReq(ctx, t, res)
-	res.Create(ctx, req, resp)
-
-	if resp.Diagnostics.HasError() {
-		t.Errorf("DBaaS User Create() reported error: %v", resp.Diagnostics)
-	}
-}
-
-// TestDBaaSUserUpdate_APIError verifies that DBaaS user Update() adds an error
-// when the initial GET returns 500.  This increases Update coverage beyond the
-// basic APIError test which may share the same code path.
-func TestDBaaSUserUpdate_APIError(t *testing.T) {
-	ctx := context.Background()
-
-	_, mockClient := newMockArubaClient(t, func(w http.ResponseWriter, r *http.Request) {
-		apiError(w, http.StatusInternalServerError)
-	})
-
-	res := NewDBaaSUserResource()
-	configureResource(ctx, t, res, mockClient)
-
-	req, resp := resourceUpdateReqFull(ctx, t, res)
-	res.Update(ctx, req, resp)
-
-	if !resp.Diagnostics.HasError() {
-		t.Error("DBaaS User Update() should fail for HTTP 500 response")
-	}
-}
-
-// TestDatabaseUpdate_Success verifies that database Update() succeeds when the
-// API returns valid data.  The Update needs projectID and dbaasID from state,
-// and the GET response needs the current database details.
-func TestDatabaseUpdate_Success(t *testing.T) {
-	ctx := context.Background()
-
-	_, mockClient := newMockArubaClient(t, func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte(`{"name":"testdb","status":{"state":"Active"}}`)) //nolint:errcheck
-	})
-
-	res := NewDatabaseResource()
-	configureResource(ctx, t, res, mockClient)
-
-	req, resp := resourceUpdateReq(ctx, t, res)
-	res.Update(ctx, req, resp)
-
-	if resp.Diagnostics.HasError() {
-		t.Errorf("Database Update() reported error: %v", resp.Diagnostics)
-	}
-}
-
-// TestProjectCreate_Success exercises the Create() path that is currently
-// low-coverage because the project API response doesn't use WaitForResourceActive.
-func TestProjectCreate_SuccessWithProperties(t *testing.T) {
-	ctx := context.Background()
-
-	projectJSON := `{"metadata":{"id":"test-id","name":"test-name","uri":"/projects/test-id"},"status":{"state":"Active"},"properties":{"description":"test desc"}}`
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		if r.Method == http.MethodPost {
-			w.WriteHeader(http.StatusCreated)
-		} else {
-			w.WriteHeader(http.StatusOK)
-		}
-		w.Write([]byte(projectJSON)) //nolint:errcheck
-	}
-
-	_, mockClient := newMockArubaClient(t, handler)
-
-	res := NewProjectResource()
-	configureResource(ctx, t, res, mockClient)
-
-	req, resp := resourceCreateReq(ctx, t, res)
-	res.Create(ctx, req, resp)
-
-	if resp.Diagnostics.HasError() {
-		t.Errorf("Project Create() reported error with properties: %v", resp.Diagnostics)
-	}
-}
-
-// TestProjectUpdate_WithProperties covers the project Update path.
-func TestProjectUpdate_WithProperties(t *testing.T) {
-	ctx := context.Background()
-
-	projectJSON := `{"metadata":{"id":"test-id","name":"test-name","location":{"value":"test-loc"}},"status":{"state":"Active"},"properties":{"description":"test desc"}}`
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte(projectJSON)) //nolint:errcheck
-	}
-
-	_, mockClient := newMockArubaClient(t, handler)
-
-	res := NewProjectResource()
-	configureResource(ctx, t, res, mockClient)
-
-	req, resp := resourceUpdateReq(ctx, t, res)
-	res.Update(ctx, req, resp)
-
-	if resp.Diagnostics.HasError() {
-		t.Errorf("Project Update() reported error with properties: %v", resp.Diagnostics)
-	}
-}
-
-// TestDatabaseGrantRead_WithProperties covers the Read() path for databasegrant
-// with a non-empty API response that includes the grant-specific fields.
-func TestDatabaseGrantRead_WithProperties(t *testing.T) {
-	ctx := context.Background()
-
-	grantJSON := `{"metadata":{"id":"test-id","name":"testgrant"},"status":{"state":"Active"},"properties":{"grantee":{"id":"test-user"},"database":{"name":"testdb"}}}`
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		if r.Method == http.MethodGet {
-			w.Header().Set("Content-Type", "application/json")
-			w.WriteHeader(http.StatusOK)
-			w.Write([]byte(grantJSON)) //nolint:errcheck
-			return
-		}
-		apiError(w, http.StatusInternalServerError)
-	}
-
-	_, mockClient := newMockArubaClient(t, handler)
-
-	res := NewDatabaseGrantResource()
-	configureResource(ctx, t, res, mockClient)
-
-	req, resp := resourceReadReq(ctx, t, res)
-	res.Read(ctx, req, resp)
-
-	// Don't assert error/no-error since the response format may not match
-	// the SDK's exact field mapping; what matters is that the code runs.
-	_ = resp
-}
-
-// TestElasticIPUpdate_WithProperties covers the elasticip Update() path with
-// a response that includes properties so property-mapping branches are covered.
-func TestElasticIPUpdate_WithProperties(t *testing.T) {
-	ctx := context.Background()
-
-	elasticipJSON := `{"metadata":{"id":"test-id","name":"test-name","location":{"value":"test-loc"}},"status":{"state":"Active"},"properties":{"address":"10.0.0.1","billingPeriod":"Hour"}}`
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte(elasticipJSON)) //nolint:errcheck
-	}
-
-	_, mockClient := newMockArubaClient(t, handler)
-
-	res := NewElasticIPResource()
-	configureResource(ctx, t, res, mockClient)
-
-	req, resp := resourceUpdateReq(ctx, t, res)
-	res.Update(ctx, req, resp)
-
-	// Don't assert success/error since response format may differ from SDK expectations
-	_ = resp
-}
-
-// TestProjectRead_WithProperties covers project Read() with a response that
-// includes properties.description so the property-mapping branch is covered.
-func TestProjectRead_WithProperties(t *testing.T) {
-	ctx := context.Background()
-
-	projectJSON := `{"metadata":{"id":"test-id","name":"test-name"},"status":{"state":"Active"},"properties":{"description":"test description"}}`
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		if r.Method == http.MethodGet {
-			w.Header().Set("Content-Type", "application/json")
-			w.WriteHeader(http.StatusOK)
-			w.Write([]byte(projectJSON)) //nolint:errcheck
-			return
-		}
-		apiError(w, http.StatusInternalServerError)
-	}
-
-	_, mockClient := newMockArubaClient(t, handler)
-
-	res := NewProjectResource()
-	configureResource(ctx, t, res, mockClient)
-
-	req, resp := resourceReadReq(ctx, t, res)
-	res.Read(ctx, req, resp)
-
-	// Property-mapping branches are covered regardless of whether the response
-	// format exactly matches the SDK's expectations.
-	_ = resp
-}
-
-// TestKMSRead_WithBillingPeriod covers the kms.Properties.BillingPeriod branch
-// that is only reached when the properties block in the API response has a
-// non-empty billingPeriod field.
-func TestKMSRead_WithBillingPeriod(t *testing.T) {
-	ctx := context.Background()
-
-	kmsJSON := `{"metadata":{"id":"test-id","name":"test-name"},"status":{"state":"Active"},"properties":{"billingPeriod":"Hour"}}`
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		if r.Method == http.MethodGet {
-			w.Header().Set("Content-Type", "application/json")
-			w.WriteHeader(http.StatusOK)
-			w.Write([]byte(kmsJSON)) //nolint:errcheck
-			return
-		}
-		apiError(w, http.StatusInternalServerError)
-	}
-
-	_, mockClient := newMockArubaClient(t, handler)
-
-	res := NewKMSResource()
-	configureResource(ctx, t, res, mockClient)
-
-	req, resp := resourceReadReq(ctx, t, res)
-	res.Read(ctx, req, resp)
-
-	if resp.Diagnostics.HasError() {
-		t.Errorf("KMS Read() reported error with billingPeriod: %v", resp.Diagnostics)
-	}
-}
-
-// TestSecurityGroupRead_WithRuleCount covers securitygroup Read() with a
-// response that includes a properties block.
-func TestSecurityGroupRead_WithProperties(t *testing.T) {
-	ctx := context.Background()
-
-	sgJSON := `{"metadata":{"id":"test-id","name":"test-name"},"status":{"state":"Active"},"properties":{"description":"test-sg","rulesCount":2}}`
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		if r.Method == http.MethodGet {
-			w.Header().Set("Content-Type", "application/json")
-			w.WriteHeader(http.StatusOK)
-			w.Write([]byte(sgJSON)) //nolint:errcheck
-			return
-		}
-		apiError(w, http.StatusInternalServerError)
-	}
-
-	_, mockClient := newMockArubaClient(t, handler)
-
-	res := NewSecurityGroupResource()
-	configureResource(ctx, t, res, mockClient)
-
-	req, resp := resourceReadReq(ctx, t, res)
-	res.Read(ctx, req, resp)
-
-	if resp.Diagnostics.HasError() {
-		t.Errorf("SecurityGroup Read() reported error with properties: %v", resp.Diagnostics)
-	}
-}
-
-// TestDBaaSUserRead_WithUsername covers dbaasuser Read() with a response that
-// includes the username field so data.Username is set from the API.
-func TestDBaaSUserRead_WithUsername(t *testing.T) {
-	ctx := context.Background()
-
-	userJSON := `{"username":"test-user","status":{"state":"Active"}}`
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		if r.Method == http.MethodGet {
-			w.Header().Set("Content-Type", "application/json")
-			w.WriteHeader(http.StatusOK)
-			w.Write([]byte(userJSON)) //nolint:errcheck
-			return
-		}
-		apiError(w, http.StatusInternalServerError)
-	}
-
-	_, mockClient := newMockArubaClient(t, handler)
-
-	res := NewDBaaSUserResource()
-	configureResource(ctx, t, res, mockClient)
-
-	req, resp := resourceReadReq(ctx, t, res)
-	res.Read(ctx, req, resp)
-
-	// The response format may or may not match exactly; what matters is coverage.
-	_ = resp
-}
-
-// TestSnapshotRead_WithProperties covers snapshot Read() with properties so
-// the volume-URI branch and billingPeriod branch are exercised.
-func TestSnapshotRead_WithProperties(t *testing.T) {
-	ctx := context.Background()
-
-	snapJSON := `{"metadata":{"id":"test-id","name":"test-name"},"status":{"state":"Active"},` +
-		`"properties":{"volume":{"uri":"/test/vol-id"},"billingPeriod":"Hour"}}`
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		if r.Method == http.MethodGet {
-			w.Header().Set("Content-Type", "application/json")
-			w.WriteHeader(http.StatusOK)
-			w.Write([]byte(snapJSON)) //nolint:errcheck
-			return
-		}
-		apiError(w, http.StatusInternalServerError)
-	}
-
-	_, mockClient := newMockArubaClient(t, handler)
-
-	res := NewSnapshotResource()
-	configureResource(ctx, t, res, mockClient)
-
-	req, resp := resourceReadReq(ctx, t, res)
-	res.Read(ctx, req, resp)
-
-	if resp.Diagnostics.HasError() {
-		t.Errorf("Snapshot Read() reported error with properties: %v", resp.Diagnostics)
-	}
-}
-
-// TestBlockStorageRead_WithProperties covers blockstorage Read() with a
-// properties block that includes bootable=true and a non-empty image field.
-func TestBlockStorageRead_WithProperties(t *testing.T) {
-	ctx := context.Background()
-
-	bsJSON := `{"metadata":{"id":"test-id","name":"test-name"},"status":{"state":"Active"},` +
-		`"properties":{"sizeGB":50,"billingPeriod":"Hour","type":"Standard","zone":"test-zone","bootable":true,"image":"test-image"}}`
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		if r.Method == http.MethodGet {
-			w.Header().Set("Content-Type", "application/json")
-			w.WriteHeader(http.StatusOK)
-			w.Write([]byte(bsJSON)) //nolint:errcheck
-			return
-		}
-		apiError(w, http.StatusInternalServerError)
-	}
-
-	_, mockClient := newMockArubaClient(t, handler)
-
-	res := NewBlockStorageResource()
-	configureResource(ctx, t, res, mockClient)
-
-	req, resp := resourceReadReq(ctx, t, res)
-	res.Read(ctx, req, resp)
-
-	if resp.Diagnostics.HasError() {
-		t.Errorf("BlockStorage Read() reported error with properties: %v", resp.Diagnostics)
-	}
-}
-
-// TestBackupRead_WithProperties covers backup Read() with a properties block
-// that includes retentionDays and billingPeriod to cover those pointer branches.
-func TestBackupRead_WithProperties(t *testing.T) {
-	ctx := context.Background()
-
-	backupJSON := `{"metadata":{"id":"test-id","name":"test-name"},"status":{"state":"Active"},` +
-		`"properties":{"retentionDays":7,"billingPeriod":"Month","origin":{"uri":"/volumes/test-vol"}}}`
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		if r.Method == http.MethodGet {
-			w.Header().Set("Content-Type", "application/json")
-			w.WriteHeader(http.StatusOK)
-			w.Write([]byte(backupJSON)) //nolint:errcheck
-			return
-		}
-		apiError(w, http.StatusInternalServerError)
-	}
-
-	_, mockClient := newMockArubaClient(t, handler)
-
-	res := NewBackupResource()
-	configureResource(ctx, t, res, mockClient)
-
-	req, resp := resourceReadReq(ctx, t, res)
-	res.Read(ctx, req, resp)
-
-	if resp.Diagnostics.HasError() {
-		t.Errorf("Backup Read() reported error with properties: %v", resp.Diagnostics)
-	}
-}
-
-// TestElasticIPRead_WithProperties covers elasticip Read() with a properties
-// block that includes address and billingPeriod fields.
-func TestElasticIPRead_WithProperties(t *testing.T) {
-	ctx := context.Background()
-
-	eipJSON := `{"metadata":{"id":"test-id","name":"test-name"},"status":{"state":"Active"},` +
-		`"properties":{"address":"10.0.0.1","billingPeriod":"Hour"}}`
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		if r.Method == http.MethodGet {
-			w.Header().Set("Content-Type", "application/json")
-			w.WriteHeader(http.StatusOK)
-			w.Write([]byte(eipJSON)) //nolint:errcheck
-			return
-		}
-		apiError(w, http.StatusInternalServerError)
-	}
-
-	_, mockClient := newMockArubaClient(t, handler)
-
-	res := NewElasticIPResource()
-	configureResource(ctx, t, res, mockClient)
-
-	req, resp := resourceReadReq(ctx, t, res)
-	res.Read(ctx, req, resp)
-
-	if resp.Diagnostics.HasError() {
-		t.Errorf("ElasticIP Read() reported error with properties: %v", resp.Diagnostics)
-	}
-}
-
-// TestKeypairRead_WithProperties covers keypair Read() path where the API
-// returns a public key value in the response.
-func TestKeypairRead_WithProperties(t *testing.T) {
-	ctx := context.Background()
-
-	kpJSON := `{"metadata":{"id":"test-id","name":"test-name"},"status":{"state":"Active"},` +
-		`"properties":{"value":"ssh-rsa AAAA... test@test"}}`
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		if r.Method == http.MethodGet {
-			w.Header().Set("Content-Type", "application/json")
-			w.WriteHeader(http.StatusOK)
-			w.Write([]byte(kpJSON)) //nolint:errcheck
-			return
-		}
-		apiError(w, http.StatusInternalServerError)
-	}
-
-	_, mockClient := newMockArubaClient(t, handler)
-
-	res := NewKeypairResource()
-	configureResource(ctx, t, res, mockClient)
-
-	req, resp := resourceReadReq(ctx, t, res)
-	res.Read(ctx, req, resp)
-
-	if resp.Diagnostics.HasError() {
-		t.Errorf("Keypair Read() reported error with properties: %v", resp.Diagnostics)
-	}
-}
